docs: 📚️ update org demo
