+++ examples/x6-example-features/src/pages/stage/change-mode.tsx
