+++ examples/x6-example-features/src/pages/stage/flex-image.ts
