+++ sites/x6-sites/examples/showcase/faq/index.en.md
@@ -1,5 +1,5 @@
 ---
-title: FAQ
+title: Common Functions
 order: 20
 redirect_from:
   - /zh/examples/practices
