+++ sites/x6-sites/examples/showcase/faq/index.zh.md
@@ -1,5 +1,5 @@
 ---
-title: 常见问题
+title: 常用功能
 order: 20
 redirect_from:
   - /zh/examples/practices
