+++ sites/x6-sites/examples/showcase/practices/demo/org.ts
@@ -27,9 +27,9 @@ Graph.registerNode(
       body: {
         refWidth: '100%',
         refHeight: '100%',
-        fill: '#FFFFFF',
-        stroke: '#000000',
-        strokeWidth: 2,
+        fill: '#5F95FF',
+        stroke: '#5F95FF',
+        strokeWidth: 1,
         rx: 10,
         ry: 10,
         pointerEvents: 'visiblePainted',
@@ -43,6 +43,7 @@ Graph.registerNode(
       rank: {
         refX: 0.9,
         refY: 0.2,
+        fill: '#fff',
         fontFamily: 'Courier New',
         fontSize: 14,
         textAnchor: 'end',
@@ -51,6 +52,7 @@ Graph.registerNode(
       name: {
         refX: 0.9,
         refY: 0.6,
+        fill: '#fff',
         fontFamily: 'Courier New',
         fontSize: 14,
         fontWeight: '800',
@@ -69,8 +71,8 @@ Graph.registerEdge(
       line: {
         fill: 'none',
         strokeLinejoin: 'round',
-        strokeWidth: '2',
-        stroke: '#4b4a67',
+        strokeWidth: 2,
+        stroke: '#A2B1C3',
         sourceMarker: null,
         targetMarker: null,
       },
@@ -80,8 +82,7 @@ Graph.registerEdge(
 )
 
 const graph = new Graph({
-  container: document.getElementById('container'),
-  grid: true,
+  container: document.getElementById('container')!,
   connecting: {
     anchor: 'orth',
   },
@@ -93,31 +94,23 @@ function member(
   rank: string,
   name: string,
   image: string,
-  background: string,
-  textColor: string = '#000',
 ) {
   return graph.addNode({
     x,
     y,
     shape: 'org-node',
     attrs: {
-      body: {
-        fill: background,
-        stroke: 'none',
-      },
       avatar: {
         opacity: 0.7,
         'xlink:href': image,
       },
       rank: {
         text: rank,
-        fill: textColor,
         wordSpacing: '-5px',
         letterSpacing: 0,
       },
       name: {
         text: name,
-        fill: textColor,
         fontSize: 13,
         fontFamily: 'Arial',
         letterSpacing: 0,
@@ -129,8 +122,12 @@ function member(
 function link(source: Node, target: Node, vertices: Point.PointLike[]) {
   return graph.addEdge({
     vertices,
-    source: { cell: source },
-    target: { cell: target },
+    source: {
+      cell: source,
+    },
+    target: {
+      cell: target,
+    },
     shape: 'org-edge',
   })
 }
@@ -140,47 +137,57 @@ const male =
 const female =
   'https://gw.alipayobjects.com/mdn/rms_43231b/afts/img/A*f6hhT75YjkIAAAAAAAAAAAAAARQnAQ'
 
-const bart = member(300, 70, 'CEO', 'Bart Simpson', male, '#30d0c6')
-const homer = member(
-  90,
-  200,
-  'VP Marketing',
-  'Homer Simpson',
-  male,
-  '#7c68fd',
-  '#f1f1f1',
-)
-const marge = member(
-  300,
-  200,
-  'VP Sales',
-  'Marge Simpson',
-  female,
-  '#7c68fd',
-  '#f1f1f1',
-)
-const lisa = member(
-  500,
-  200,
-  'VP Production',
-  'Lisa Simpson',
-  female,
-  '#7c68fd',
-  '#f1f1f1',
-)
-const maggie = member(400, 350, 'Manager', 'Maggie Simpson', female, '#feb563')
-const lenny = member(190, 350, 'Manager', 'Lenny Leonard', male, '#feb563')
-const carl = member(190, 500, 'Manager', 'Carl Carlson', male, '#feb563')
+const bart = member(300, 70, 'CEO', 'Bart Simpson', male)
+const homer = member(90, 200, 'VP Marketing', 'Homer Simpson', male)
+const marge = member(300, 200, 'VP Sales', 'Marge Simpson', female)
+const lisa = member(500, 200, 'VP Production', 'Lisa Simpson', female)
+const maggie = member(400, 350, 'Manager', 'Maggie Simpson', female)
+const lenny = member(190, 350, 'Manager', 'Lenny Leonard', male)
+const carl = member(190, 500, 'Manager', 'Carl Carlson', male)
 
-link(bart, marge, [{ x: 385, y: 180 }])
+link(bart, marge, [
+  {
+    x: 385,
+    y: 180,
+  },
+])
 link(bart, homer, [
-  { x: 385, y: 180 },
-  { x: 175, y: 180 },
+  {
+    x: 385,
+    y: 180,
+  },
+  {
+    x: 175,
+    y: 180,
+  },
 ])
 link(bart, lisa, [
-  { x: 385, y: 180 },
-  { x: 585, y: 180 },
+  {
+    x: 385,
+    y: 180,
+  },
+  {
+    x: 585,
+    y: 180,
+  },
 ])
-link(homer, lenny, [{ x: 175, y: 380 }])
-link(homer, carl, [{ x: 175, y: 530 }])
-link(marge, maggie, [{ x: 385, y: 380 }])
+link(homer, lenny, [
+  {
+    x: 175,
+    y: 380,
+  },
+])
+link(homer, carl, [
+  {
+    x: 175,
+    y: 530,
+  },
+])
+link(marge, maggie, [
+  {
+    x: 385,
+    y: 380,
+  },
+])
+
+graph.zoomToFit({ padding: 20, maxScale: 1 })
